CI pipeline w/ Jacoco code coverage and sonarcloud.io integration
- uses testcontainers(.org) to start ehrbase & postgresql containers w/ docker compose during mvn verify phase
- uses ehrbaseorg/ehrbase:test_0.13.0 docker image
