+++ pom.xml
@@ -25,6 +25,7 @@
         <slf4j.version>1.7.30</slf4j.version>
         <spring-boot.version>2.2.6.RELEASE</spring-boot.version>
         <woodstox.version>6.1.1</woodstox.version>
+        <ehrbase.sdk.version>develop-SNAPSHOT</ehrbase.sdk.version>
     </properties>
 
     <build>
@@ -33,17 +34,67 @@
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <version>${spring-boot.version}</version>
+                <configuration>
+                    <jvmArguments>${argLine}</jvmArguments>
+                </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-failsafe-plugin</artifactId>
-                <version>2.22.2</version>
+                <version>2.22.0</version>
                 <executions>
                     <execution>
                         <goals>
                             <goal>integration-test</goal>
                             <goal>verify</goal>
                         </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
+                <groupId>org.jacoco</groupId>
+                <artifactId>jacoco-maven-plugin</artifactId>
+                <version>0.8.5</version>
+                <executions>
+                    <execution>
+                        <id>default-prepare-agent</id>
+                        <goals>
+                            <goal>prepare-agent</goal>
+                        </goals>
+                    </execution>
+                    <execution>
+                        <id>default-report</id>
+                        <phase>verify</phase>
+                        <goals>
+                            <goal>report</goal>
+                        </goals>
+                    </execution>
+
+                    <!-- FAIL BUILD IF RULES NOT MET! -->
+                    <execution>
+                        <id>default-check</id>
+                        <goals>
+                            <goal>check</goal>
+                        </goals>
+                        <configuration>
+                            <rules>
+                                <rule>
+                                    <element>BUNDLE</element>
+                                    <limits>
+                                        <limit>
+                                            <counter>INSTRUCTION</counter>
+                                            <value>COVEREDRATIO</value>
+                                            <minimum>0.50</minimum>
+                                        </limit>
+                                        <!-- <limit>
+                                            <counter>CLASS</counter>
+                                            <value>MISSEDCOUNT</value>
+                                            <maximum>0</maximum>
+                                        </limit> -->
+                                    </limits>
+                                </rule>
+                            </rules>
+                        </configuration>
                     </execution>
                 </executions>
             </plugin>
@@ -153,25 +204,18 @@
             <artifactId>thymeleaf</artifactId>
             <version>3.0.11.RELEASE</version>
         </dependency>
-        <!--        <dependency>-->
-        <!--            <groupId>org.ehrbase</groupId>-->
-        <!--            <artifactId>client-library</artifactId>-->
-        <!--            <version>0.3.0</version>-->
-        <!--        </dependency>-->
-        <!-- this is a local installation of the dependency -->
+        <dependency>
+            <groupId>com.google.code.gson</groupId>
+            <artifactId>gson</artifactId>
+            <version>2.8.6</version>
+        </dependency>
+
+        <!-- SDK -->
         <dependency>
             <groupId>com.github.ehrbase.openEHR_SDK</groupId>
             <artifactId>client</artifactId>
-            <version>0.3.5</version>
-            <!--<version>${ehrbase.sdk.version}</version>-->
+            <version>${ehrbase.sdk.version}</version>
         </dependency>
-
-        <!--        <dependency>-->
-        <!--            <groupId>com.github.ehrbase.openEHR_SDK</groupId>-->
-        <!--            <artifactId>client-library</artifactId>-->
-        <!--            <version>d97d209</version>&lt;!&ndash; FIXME: THIS IS A SPECIFIC COMMIT &ndash;&gt;-->
-        <!--&lt;!&ndash;            <version>v0.3.4</version> This gets back the reflections errors&ndash;&gt;-->
-        <!--        </dependency>-->
 
         <!-- Test Dependencies -->
         <dependency>
@@ -186,7 +230,6 @@
             <version>5.6.2</version>
             <scope>test</scope>
         </dependency>
-
         <dependency>
             <groupId>org.testcontainers</groupId>
             <artifactId>testcontainers</artifactId>
@@ -209,12 +252,11 @@
                     <groupId>com.vaadin.external.google</groupId>
                     <artifactId>android-json</artifactId>
                 </exclusion>
+                <exclusion>
+                    <groupId>org.junit.vintage</groupId>
+                    <artifactId>junit-vintage-engine</artifactId>
+                </exclusion>
             </exclusions>
-        </dependency>
-        <dependency>
-            <groupId>com.google.code.gson</groupId>
-            <artifactId>gson</artifactId>
-            <version>2.8.6</version>
         </dependency>
     </dependencies>
 
@@ -299,10 +341,10 @@
         </dependencies>
     </dependencyManagement>
     <repositories>
-        <!--        <repository>-->
-        <!--            <id>jitpack.io</id>-->
-        <!--            <url>https://jitpack.io</url>-->
-        <!--        </repository>-->
+        <repository>
+            <id>jitpack.io</id>
+            <url>https://jitpack.io</url>
+        </repository>
         <repository>
             <id>oss-snapshots</id>
             <snapshots>
