+++ sonar-project.properties
@@ -1,0 +1,46 @@
+#
+# Copyright (c) 2020 Wladislaw Wagner (Vitasystems GmbH) and Hannover Medical School.
+#
+# This file is part of project EHRBase
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+sonar.organization=ehrbase
+sonar.projectKey=ehrbase_fhir-bridge
+# sonar.host.url=https://sonarcloud.io
+
+sonar.projectName=fhir-bridge
+sonar.projectVersion=1.0
+
+sonar.coverage.jacoco.xmlReportPaths=target/site/jacoco/jacoco.xml
+
+# Comma-separated paths to folders containing the *.xml JUnit reports. 
+sonar.junit.reportPaths=/home/circleci/test-results/integration-tests/  #, \
+    # /home/circleci/test-results/unit-tests/
+    
+
+
+sonar.java.binaries=target/**
+
+# Patterns used to exclude some files from coverage report.
+sonar.coverage.exclusions=**/test/**, mappings/**/*
+
+# Patterns used to exclude some source files from analysis.
+sonar.exclusions=doc/**, mappings/**
+
+# Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
+#sonar.sources=.
+ 
+# Encoding of the source code. Default is default system encoding
+#sonar.sourceEncoding=UTF-8
