+++ src/test/resources/ehrbase-compose.yml
@@ -11,9 +11,6 @@ services:
       DB_URL: jdbc:postgresql://ehrdb:5432/ehrbase
       DB_USER: ehrbase
       DB_PASS: ehrbase
-      # AUTH_TYPE: BASIC
-      # AUTH_USER: ehrbase-user
-      # AUTH_PASSWORD: SuperSecretPassword
       SYSTEM_NAME: local.ehrbase.org
     restart: on-failure
 
@@ -23,8 +20,6 @@ services:
       - 5432:5432
     networks:
       - ehrbase-net
-    volumes:
-      - ./.pgdata:/var/lib/postgresql/data
     environment:
       POSTGRES_USER: postgres
       POSTGRES_PASSWORD: postgres
